+++ index.html
@@ -1,15 +1,25 @@
 <!DOCTYPE html>
 <html lang="ko">
+  <head>
+    <meta charset="utf-8" />
+    <title>점심 뭐 먹지</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+  </head>
+  <body>
+    <!-- GNB -->
+    <header-element></header-element>
 
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>점심 뭐 먹지</title>
-</head>
+    <main>
+      <!-- 카테고리/정렬 필터 -->
+      <select-element></select-element>
 
-<body>
+      <!-- 음식점 목록 -->
+      <section class="restaurant-list-container">
+        <restaurant-list></restaurant-list>
+      </section>
 
-</body>
-
+      <!-- 음식점 추가 모달 -->
+      <modal-element></modal-element>
+    </main>
+  </body>
 </html>
