+++ src/components/modal/RestaurantAddFormComponent.js
@@ -1,7 +1,6 @@
-import CustomElement from "../abstracts/CustomElement";
-import dispatcher from "../domain/Dispatcher";
-import { RESTAURANT_ACTION } from "../abstracts/constants";
-import { Restaurant } from "../abstracts/types";
+import CustomElement from "../../abstracts/CustomElement";
+import dispatcher from "../../domain/Dispatcher";
+import { RESTAURANT_ACTION } from "../../abstracts/constants";
 
 class RestaurantAddFormComponent extends CustomElement {
   setEvent() {
