--- templates/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-  <head>
-    <meta charset="utf-8" />
-    <title>점심 뭐 먹지</title>
-    <link rel="stylesheet" href="style.css" />
-    <link rel="component" href="../src/components/RestaurantComponent.js" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-  </head>
-
-  <body>
-    <!-- GNB -->
-    <header-element></header-element>
-
-    <main>
-      <!-- 카테고리/정렬 필터 -->
-      <select-element></select-element>
-
-      <!-- 음식점 목록 -->
-      <section class="restaurant-list-container">
-        <restaurant-list></restaurant-list>
-      </section>
-
-      <!-- 음식점 추가 모달 -->
-      <modal-element></modal-element>
-    </main>
-  </body>
-</html>
