+++ webpack.config.js
@@ -36,7 +36,7 @@ module.exports = {
   },
   plugins: [
     new HtmlWebpackPlugin({
-      template: "./templates/index.html",
+      template: "./index.html",
     }),
   ],
   devtool: "inline-source-map",
